backend/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.10.0
psycopg2
djangorestframework-jwt
cryptography
python-jose
gunicorn==20.1.0
Pillow
coverage
